pyproject.toml
[project]
name = "dissertation"
version = "0.1.0"
description = "Dissertation scripts and data analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Data analysis
    "numpy>=2.0",
    "pandas>=2.0",
    "polars>=1.0",
    "pyarrow>=15.0",
    "scipy>=1.10",
    # Visualization
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "matplot2tikz>=0.5",
    # LaTeX
    "pylatex>=1.4",
    # Utilities
    "tabulate>=0.9",
    "pyyaml>=6.0",
    "requests>=2.28",
    # NLP/ML (for dblp_stats.py --discover)
    "scikit-learn>=1.3",
    "sentence-transformers>=5.2.1",
    "numba>=0.58",  # JIT compiler, 0.58+ for Python 3.12
    "hdbscan>=0.8.33",  # Better clustering for embedding method
    "umap-learn>=0.5.7",  # Dimensionality reduction for embedding method
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "mypy>=1.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)
